x64 support : some long changed to int.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@449 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ sources/libhxcfe.h
@@ -240,7 +240,7 @@ typedef struct s_trackdisplay_
 	int xsize,ysize;
 	int x_us,y_us;
 	int x_start_us;
-	unsigned long * framebuffer;
+	unsigned int * framebuffer;
 
 	s_sectorlist * sl;
